pyproject.toml
[project]
name = "ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "inquirerpy>=0.3.4",
    "keyring>=25.6.0",
    "openai>=1.60.2",
    "python-dotenv>=1.0.1",
]
